backend/pyproject.toml
[project]
name = "retrospekt-backend"
version = "1.9.0"
description = "Backend for Retrospekt retro board"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "motor>=3.6.0",
    "pydantic-settings>=2.0.0",
    "sse-starlette>=2.1.0",
    "redis>=5.0.0",
    "sentry-sdk[fastapi]>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "mongomock-motor>=0.0.36",
    "httpx>=0.28.0",
    "nox>=2024.10.9",
    "ruff>=0.8.0",
    "mypy>=1.9.0",
    "fakeredis>=2.20.0",
]

[project.scripts]
retrospekt-server = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["src/main.py"]  # uvicorn entry point — covered by E2E

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 85

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
